Revise BarChart styling and sizing.
Putting a ResponsiveContainer inside this component is problematic. If a consumer of the component doesn't apply styles correctly, either to the containing element or via style props, the component is likely to render with height and/or width of 0, making the chart invisible. The revised interface requires specifying the chart dimensions (with defaults applied if not specified). While this approach offers a bit less flexibility, I think it makes the sizing easier to understand and consumption of the component far less error-prone.